+++ src/apps/weblib/js-api/vizzu.js
@@ -409,11 +409,6 @@ export default class Vizzu {
     }
   }
 
-  _getPointerPos(evt) {
-    const rect = this.render.clientRect()
-    return [evt.clientX - rect.left, evt.clientY - rect.top]
-  }
-
   _toCString(str) {
     const len = str.length * 4 + 1
     const buffer = this.module._malloc(len)
@@ -475,23 +470,19 @@ export default class Vizzu {
 
     this._resizeObserver.observe(canvas)
 
-    this._resizeHandler = () => {
-      this.render.updateFrame(true)
-    }
-
     this._pointermoveHandler = (evt) => {
-      const pos = this._getPointerPos(evt)
-      this._call(this.module._vizzu_pointerMove)(evt.pointerId, pos[0], pos[1])
+      const pos = this.render.clientToRenderCoor({ x: evt.clientX, y: evt.clientY })
+      this._call(this.module._vizzu_pointerMove)(evt.pointerId, pos.x, pos.y)
     }
 
     this._pointerupHandler = (evt) => {
-      const pos = this._getPointerPos(evt)
-      this._call(this.module._vizzu_pointerUp)(evt.pointerId, pos[0], pos[1])
+      const pos = this.render.clientToRenderCoor({ x: evt.clientX, y: evt.clientY })
+      this._call(this.module._vizzu_pointerUp)(evt.pointerId, pos.x, pos.y)
     }
 
     this._pointerdownHandler = (evt) => {
-      const pos = this._getPointerPos(evt)
-      this._call(this.module._vizzu_pointerDown)(evt.pointerId, pos[0], pos[1])
+      const pos = this.render.clientToRenderCoor({ x: evt.clientX, y: evt.clientY })
+      this._call(this.module._vizzu_pointerDown)(evt.pointerId, pos.x, pos.y)
     }
 
     this._pointerleaveHandler = (evt) => {
@@ -515,7 +506,6 @@ export default class Vizzu {
       }
     }
 
-    window.addEventListener('resize', this._resizeHandler)
     canvas.addEventListener('pointermove', this._pointermoveHandler)
     canvas.addEventListener('pointerup', this._pointerupHandler)
     canvas.addEventListener('pointerdown', this._pointerdownHandler)
@@ -528,7 +518,6 @@ export default class Vizzu {
     this?._resizeObserver.disconnect()
     if (this._pollInterval) clearInterval(this._pollInterval)
     if (this._updateInterval) clearInterval(this._updateInterval)
-    if (this._resizeHandler) window.removeEventListener('resize', this._resizeHandler)
     if (this._pointermoveHandler)
       this?.canvas.removeEventListener('pointermove', this._pointermoveHandler)
     if (this._pointerupHandler)