+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased] - 2023-08-24
+## [Unreleased]
 
 ### Fixed
 
@@ -16,6 +16,7 @@
   chart dimension axis ranges.
 - When measure set on color channel, the color changes alongside the color range
   instead of blending the starting and ending colors.
+- Pointer event coordinates fixed when canvas is in CSS transformed view.
 
 ### Added